Got AUTO working
READY
AUTO 10
10 FOR A# = 1 TO 5
20 PRINT A#
30 NEXT A#
40 LIST
10 FOR A# = 1 TO 5
20 PRINT A#
30 NEXT A#
40 RUN
1
2
3
4
5
READY
AUTO 10
40 PRINT "LOOP DONE"
50 AUTO
LIST
10 FOR A# = 1 TO 5
20 PRINT A#
30 NEXT A#
40 PRINT "LOOP DONE"
RUN
1
2
3
4
5
LOOP DONE
READY

+++ basicenvironment.go
@@ -11,9 +11,9 @@ type BasicEnvironment struct {
 	forStepValue BasicValue
 	forToLeaf *BasicASTLeaf
 	forToValue BasicValue
-	loopFirstLine int
-	loopExitLine int
-	gosubReturnLine int
+	loopFirstLine int64
+	loopExitLine int64
+	gosubReturnLine int64
 	runtime *BasicRuntime
 }
